+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [1.3.0] - 2019-10-17
+
+- Added
+  - Additional webhook properties: filter, hookAttribute
+  - Deleting buckets
+
 ## [1.2.0] - 2019-10-09
 
 - Added